events_ring: bail-out on prev_busy + more counters
If several writers are active, even if the write index is
atomically increased, it is still possible for 2 writers to try to
write the same entry, if the write index wraps around.
E.g.: event_ring with of size 4m index 0 and while writing
event[0] in one thread, 4 more threads try to write => the 4th one
would get the same event ring destination (4 mod 4 = 0).
In this case just give-up (write fails/event dropped): writers are
just to fast for the event buffer size.