tcp process packets: use low reorder wait timeout on start-up
On start-up it's very probable to see packets belonging to ongoing
tcp connections (connections for which started before). In this
case we don't want to wait a long time for possible out-of-order
packet (the 1st packet we see will look like out-of-order, but we
can't actually tell if it really is or we just missed the
connection establishment phase because we were not running at the
time).
To mitigate this, the first 60s after start-up use a 1s
reorder_wait timeout and a 1s tcp garbage collection timeout
(these values are hard-wired for now).
After that revert to the values from the config.