evrate: support for reporting exceeded count diff
Add to each blacklisted event the exceeded count difference
since the last report for the same event:ip (last added event marked
as blacklisted).
Note that the repeat/diff count is deduced and so it will be accurate
only if the blacklist report parameters are not changed in the
meantime (evr_conseq_report_min and evr_conseq_report_max).