disable promiscuous mode for the "any" interface
When configured to listen on all interfaces (interface "any"),
promiscuous mode is not supported and it will be automatically
disabled.
This avoids a bug in gopacket (it treats libpcap warnings as
errors and is triggered by newer libpcap versions that signal a
warning when trying to listen on "any" in promiscuous mode).